projects/Gemini Assistant (Shared)/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
python-multipart==0.0.20
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests
tavily-python
